Refactor metrics and logging integration across services
- Simplified metric creation for script execution and command processing.
- Updated event handlers to use command names for better tracking.
- Consolidated service registration for settings providers, utilities, and Discord components.
- Enhanced logging capabilities with structured information for build metadata.
- Removed redundant code and improved clarity in service initialization.
- Introduced a metrics server setup based on global settings.
